# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ilter Advanced Gateway API
  description: Secure, cost-optimizing LLM Gateway with built-in PII masking, loop detection, and load balancing.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:9191
    description: Dashboard Server
paths:
...
  /api/cache/flush:
    post:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Flush Semantic Cache
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cache flushed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: flushed
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/chat/completions:
    post:
      tags: [Proxy]
      servers:
        - url: http://127.0.0.1:8181
      summary: Proxy Chat Completion
      description: Proxies chat completions to the load-balanced providers.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                model:
                  type: string
                  example: gpt-4o
                messages:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant, tool]
                        example: user
                      content:
                        type: string
                        example: Hello, how are you?
                stream:
                  type: boolean
                  example: false
                temperature:
                  type: number
                  minimum: 0.0
                  maximum: 2.0
                  example: 0.7
                max_tokens:
                  type: integer
                  minimum: 1
                  example: 1024
      responses:
        '200':
          description: Completion successful
          headers:
            X-Ilter-Complexity-Score:
              schema:
                type: number
                description: Complexity score of the prompt (0.0–10.0)
            X-Ilter-Model-Selected:
              schema:
                type: string
                description: The model selected by smart routing
            X-Ilter-Cost:
              schema:
                type: number
                description: Actual cost in USD for this request (non-streaming only)
            X-Ilter-Cost-Estimate:
              schema:
                type: number
                description: Estimated cost in USD for this request
            X-Ilter-Alternative-Cost:
              schema:
                type: number
                description: Estimated cost if using an alternative model
            X-Ilter-Savings-Potential:
              schema:
                type: string
                description: Savings percentage compared to alternative
                example: 45%
            X-Ilter-Loop-Warning:
              schema:
                type: boolean
                description: Indicates potential agentic loop detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  object:
                    type: string
                  created:
                    type: integer
                  model:
                    type: string
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        message:
                          type: object
                          properties:
                            role:
                              type: string
                            content:
                              type: string
                        finish_reason:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/PIIBlocked'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/ProviderError'
  /v1/models:
    get:
      tags: [Proxy]
      servers:
        - url: http://127.0.0.1:8181
      summary: List Models
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of available models
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: list
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        object:
                          type: string
                        created:
                          type: integer
                        owned_by:
                          type: string
                        provider:
                          type: string
                        type:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/stats:
    get:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Get Topline Dashboard Statistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_requests:
                    type: integer
                  total_cost:
                    type: number
                  estimated_savings:
                    type: number
                  success_count:
                    type: integer
                  error_count:
                    type: integer
                  cache_hits:
                    type: integer
                  avg_latency_ms:
                    type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/requests:
    get:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Get Audit Logs list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [success, error]
          description: Filter by response status
        - name: model
          in: query
          required: false
          schema:
            type: string
          description: Filter by model name (LIKE match)
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        timestamp:
                          type: string
                          format: date-time
                        api_key_id:
                          type: integer
                        model:
                          type: string
                        provider:
                          type: string
                        prompt_tokens:
                          type: integer
                        completion_tokens:
                          type: integer
                        cached_tokens:
                          type: integer
                          description: Cache-read (cached input) token count
                        cache_creation_tokens:
                          type: integer
                          description: Anthropic cache-write token count
                        total_cost:
                          type: number
                        latency_ms:
                          type: integer
                        status_code:
                          type: integer
                        cache_hit:
                          type: boolean
                        prompt_preview:
                          type: string
                        request_body:
                          type: string
                        response_body:
                          type: string
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/costs:
    get:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Get Cost Attribution with Time Series and Savings
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: ["24h", "7d", "30d"]
            default: "7d"
          description: Time period for cost breakdown (24h, 7d, 30d)
      responses:
        '200':
          description: Cost attribution retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_cost:
                    type: number
                    format: double
                  total_requests:
                    type: integer
                  avg_cost_per_request:
                    type: number
                    format: double
                  period:
                    type: string
                  by_provider:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                        cost:
                          type: number
                          format: double
                        count:
                          type: integer
                        pct:
                          type: number
                          format: double
                  time_series:
                    type: array
                    items:
                      type: object
                      properties:
                        period:
                          type: string
                        cost:
                          type: number
                          format: double
                        count:
                          type: integer
                  savings_summary:
                    type: object
                    properties:
                      routing_savings:
                        type: number
                        format: double
                      cache_savings:
                        type: number
                        format: double
                      total_savings:
                        type: number
                        format: double
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/models:
    get:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Get Active Models List
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Models list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    provider:
                      type: string
                    type:
                      type: string
                    owned_by:
                      type: string
                    active:
                      type: boolean
                    configured:
                      type: boolean
                    display_name:
                      type: string
                    category:
                      type: string
                    cost_per_input_token:
                      type: number
                      format: float
                    cost_per_output_token:
                      type: number
                      format: float
                    cost_per_cached_input_token:
                      type: number
                      format: float
                    cost_per_cache_write_token:
                      type: number
                      format: float
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/loops:
    get:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Get Loop Events list
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Loop events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    detected_at:
                      type: string
                      format: date-time
                    api_key_id:
                      type: integer
                    client_ip:
                      type: string
                    prompt_hash:
                      type: string
                    repeat_count:
                      type: integer
                    window_seconds:
                      type: integer
                    action_taken:
                      type: string
                    resolved_at:
                      type: string
                      format: date-time
                      nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/optimize:
    post:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Optimization Model Recommendation Helper
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
                - current_model
              properties:
                prompt:
                  type: string
                  example: Say hello.
                current_model:
                  type: string
                  example: gpt-4o
      responses:
        '200':
          description: Recommendations computed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  complexity_score:
                    type: number
                  current_cost_estimate:
                    type: number
                  recommendations:
                    type: array
                    items:
                      type: object
                      properties:
                        model:
                          type: string
                        category:
                          type: string
                        input_cost:
                          type: number
                        output_cost:
                          type: number
                        savings_percent:
                          type: number
                        reason:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/models/category:
    post:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Update model category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
              properties:
                name:
                  type: string
                  example: gpt-4o
                category:
                  type: string
                  enum: [free, economy, standard, premium, embedding]
                  example: premium
      responses:
        '200':
          description: Model category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Model not found
        '400':
          description: Invalid category value
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    BadRequest:
      description: Malformed request payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    example: model_not_found
                  code:
                    type: string
                    example: "404"
...